requirements.txt
torch
pillow
transformers
aiogram